[Login] 빠른 로딩때문에 userInfo를 get하지 못함 > 4초 로딩화면추가로 해결 > 반성일기 넣는부분 확보
