+++ frontend/src/components/Navigation/UserInfo.js
@@ -9,30 +9,12 @@ import {
   userInfoWrap,
   avatar,
 } from './StyledNavigation'
+import Loader from '../../components/Loader'
 import UserInfoPopup from './UserInfoPopup'
 import { getUserInfo } from '../../actions/auth.js'
-import rootApi from '../../config'
 import { connect } from 'react-redux'
 
 class UserInfo extends Component {
-  componentDidMount() {
-    if (this.props.token) {
-      this.getUserInfo()
-    }
-  }
-  getUserInfo = () => {
-    fetch(`${rootApi}/user`, {
-      method: 'GET',
-      headers: {
-        Authorization: `Bearer ${window
-          .localStorage.token}`,
-      },
-    })
-      .then(res => res.json())
-      .then(userInfo => {
-        this.props.I_WANT_SAVE_USER_INFO(userInfo)
-      })
-  }
   render() {
     return (
       <Menu.Item style={userInfoWrap}>
@@ -70,18 +52,9 @@ class UserInfo extends Component {
 const mapStateToProps = state => {
   return {
     userInfo: state.auth.userInfo,
-    token: state.auth.token,
   }
 }
 
-const mapDispatchtoProps = dispatch => {
-  return {
-    I_WANT_SAVE_USER_INFO: user =>
-      dispatch(getUserInfo(user)),
-  }
-}
-
-export default connect(
-  mapStateToProps,
-  mapDispatchtoProps,
-)(UserInfo)
+export default connect(mapStateToProps, null)(
+  UserInfo,
+)
